refactored TestUtils class/fixture to a utils.py module.
