+++ tests/test_modules/test_detection_and_extraction_module.py
@@ -1,9 +1,9 @@
 import numpy as np
-from pathlib import Path
-from suite2p import extraction, registration
 
+from suite2p import extraction
+import utils
 
-def prepare_for_detection(op, input_file_name_list, dimensions, test_utils):
+def prepare_for_detection(op, input_file_name_list, dimensions):
     """
     Prepares for detection by filling out necessary ops parameters. Removes dependence on
     other modules. Creates pre_registered binary file.
@@ -16,15 +16,15 @@ def prepare_for_detection(op, input_file_name_list, dimensions, test_utils):
     ops = []
     for plane in range(op['nplanes']):
         curr_op = op.copy()
-        plane_dir = test_utils.get_plane_dir(op, plane)
-        bin_path = test_utils.write_data_to_binary(
+        plane_dir = utils.get_plane_dir(op, plane)
+        bin_path = utils.write_data_to_binary(
             str(plane_dir.joinpath('data.bin')), str(input_file_name_list[plane][0])
         )
         curr_op['reg_file'] = bin_path
         if plane == 1: # Second plane result has different crop.
             curr_op['xrange'], curr_op['yrange'] = [[1, 403], [1, 359]]
         if curr_op['nchannels'] == 2:
-            bin2_path = test_utils.write_data_to_binary(
+            bin2_path = utils.write_data_to_binary(
                 str(plane_dir.joinpath('data_chan2.bin')), str(input_file_name_list[plane][1])
             )
             curr_op['reg_file_chan2'] = bin2_path
@@ -34,10 +34,10 @@ def prepare_for_detection(op, input_file_name_list, dimensions, test_utils):
     return ops
 
 
-def detect_and_extract_wrapper(ops, test_utils):
+def detect_and_extract_wrapper(ops):
     for plane in range(ops[0]['nplanes']):
         curr_op = ops[plane]
-        plane_dir = test_utils.get_plane_dir(curr_op, plane)
+        plane_dir = utils.get_plane_dir(curr_op, plane)
         # Detection Part
         curr_op = extraction.detect_and_extract(curr_op)
         # Extraction part
@@ -55,13 +55,14 @@ class TestSuite2pDetectionExtractionModule:
     """
     # TODO: Separate once the detection module is created.
 
-    def test_detection_extraction_output_1plane1chan(self, default_ops, test_utils):
+    def test_detection_extraction_output_1plane1chan(self, default_ops):
         ops = prepare_for_detection(
-            default_ops, [[default_ops['data_path'][0].joinpath('detection', 'pre_registered.npy')]], (404, 360),
-            test_utils
+            default_ops,
+            [[default_ops['data_path'][0].joinpath('detection', 'pre_registered.npy')]],
+            (404, 360)
         )
-        detect_and_extract_wrapper(ops, test_utils)
-        test_utils.check_output(
+        detect_and_extract_wrapper(ops)
+        utils.check_output(
             default_ops['save_path0'],
             ['F', 'Fneu', 'iscell', 'stat', 'spks'],
             default_ops['data_path'][0],
@@ -69,7 +70,7 @@ class TestSuite2pDetectionExtractionModule:
             default_ops['nchannels'],
         )
 
-    def test_detection_extraction_output_2plane2chan(self, default_ops, test_utils):
+    def test_detection_extraction_output_2plane2chan(self, default_ops):
         default_ops['nchannels'] = 2
         default_ops['nplanes'] = 2
         detection_dir = default_ops['data_path'][0].joinpath('detection')
@@ -80,12 +81,11 @@ class TestSuite2pDetectionExtractionModule:
                 [detection_dir.joinpath('pre_registered11.npy'), detection_dir.joinpath('pre_registered12.npy')]
             ]
             , (404, 360),
-            test_utils
         )
         ops[0]['meanImg_chan2'] = np.load(detection_dir.joinpath('meanImg_chan2p0.npy'))
         ops[1]['meanImg_chan2'] = np.load(detection_dir.joinpath('meanImg_chan2p1.npy'))
-        detect_and_extract_wrapper(ops, test_utils)
-        test_utils.check_output(
+        detect_and_extract_wrapper(ops)
+        utils.check_output(
             default_ops['save_path0'],
             ['F', 'Fneu', 'iscell', 'stat', 'spks'],
             default_ops['data_path'][0],
